ANDA!

+++ Trabajo Practico 1/FuncionesAritmeticas.h
@@ -3,5 +3,5 @@ int Suma(int numero1,int numero2);
 int Resta(int numero1, int numero2);
 int Multiplicacion(int factor1,int factor2);
 float Division(float dividendo,float divisor);
-float Factorial(int primerOperando);
+float Factorial(float primerOperando);
